+++ kernel/thread.c
@@ -1613,3 +1613,30 @@ static inline void z_vrfy_k_thread_abort(k_tid_t thread)
 }
 #include <zephyr/syscalls/k_thread_abort_mrsh.c>
 #endif /* CONFIG_USERSPACE */
+
+bool k_can_yield(void)
+{
+	unsigned int k = arch_irq_lock();
+	bool irq_locked = !arch_irq_unlocked(k);
+
+	arch_irq_unlock(k);
+	return !(k_is_pre_kernel() || k_is_in_isr() || irq_locked ||
+		 z_is_idle_thread_object(_current));
+}
+
+void z_impl_k_yield(void)
+{
+	__ASSERT(!arch_is_in_isr(), "");
+
+	SYS_PORT_TRACING_FUNC(k_thread, yield);
+
+	z_sched_yield();
+}
+
+#ifdef CONFIG_USERSPACE
+static inline void z_vrfy_k_yield(void)
+{
+	z_impl_k_yield();
+}
+#include <zephyr/syscalls/k_yield_mrsh.c>
+#endif /* CONFIG_USERSPACE */
